services/analytics-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
clickhouse-connect==0.6.18
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
